Not broken; currently figuring out how to get enable/disable functionality working
