+++ src/Microsoft.Diagnostics.Monitoring.WebApi/CollectionRules/ICollectionRuleTriggerOptions.cs
@@ -6,23 +6,16 @@ using Microsoft.Diagnostics.Monitoring.WebApi;
 using System.ComponentModel.DataAnnotations;
 using System.Diagnostics;
 
-namespace Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Options
+namespace Microsoft.Diagnostics.Monitoring.WebApi
 {
     /// <summary>
     /// Options for describing the type of trigger and the settings to pass to that trigger.
     /// </summary>
-    [DebuggerDisplay("Trigger: Type = {Type}")]
-    internal sealed partial class CollectionRuleTriggerOptions
+    public class CollectionRuleTriggerOptions
     {
-        [Display(
-            ResourceType = typeof(OptionsDisplayStrings),
-            Description = nameof(OptionsDisplayStrings.DisplayAttributeDescription_CollectionRuleTriggerOptions_Type))]
         [Required]
         public string Type { get; set; }
 
-        [Display(
-            ResourceType = typeof(OptionsDisplayStrings),
-            Description = nameof(OptionsDisplayStrings.DisplayAttributeDescription_CollectionRuleTriggerOptions_Settings))]
         public object Settings { get; internal set; }
     }
 }
