+++ src/Tools/dotnet-monitor/CollectionRules/Configuration/CollectionRuleConfigureNamedOptions.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.Diagnostics.Monitoring.WebApi;
 using Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Actions;
 using Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Options;
 using Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Triggers;
@@ -52,7 +53,7 @@ namespace Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Configuration
 
         private void BindActionSettings(IConfigurationSection ruleSection, CollectionRuleOptions ruleOptions, int actionIndex)
         {
-            CollectionRuleActionOptions actionOptions = ruleOptions.Actions[actionIndex];
+            CollectionRuleActionOptions actionOptions = (CollectionRuleActionOptions)ruleOptions.Actions[actionIndex];
 
             if (null != actionOptions &&
                 _actionOperations.TryCreateOptions(actionOptions.Type, out object actionSettings))
@@ -70,7 +71,7 @@ namespace Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Configuration
 
         private void BindTriggerSettings(IConfigurationSection ruleSection, CollectionRuleOptions ruleOptions)
         {
-            CollectionRuleTriggerOptions triggerOptions = ruleOptions.Trigger;
+            CollectionRuleTriggerOptions triggerOptions = (CollectionRuleTriggerOptions)ruleOptions.Trigger;
 
             if (null != triggerOptions &&
                 _triggerOperations.TryCreateOptions(triggerOptions.Type, out object triggerSettings))
