--- src/Tools/dotnet-monitor/CollectionRules/Options/CollectionRuleLimitsOptions.cs
@@ -1,36 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using Microsoft.Diagnostics.Monitoring.WebApi;
-using System;
-using System.ComponentModel;
-using System.ComponentModel.DataAnnotations;
-
-namespace Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Options
-{
-    /// <summary>
-    /// Options for limiting the execution of a collection rule.
-    /// </summary>
-    internal sealed class CollectionRuleLimitsOptions
-    {
-        [Display(
-            ResourceType = typeof(OptionsDisplayStrings),
-            Description = nameof(OptionsDisplayStrings.DisplayAttributeDescription_CollectionRuleLimitsOptions_ActionCount))]
-        [DefaultValue(CollectionRuleLimitsOptionsDefaults.ActionCount)]
-        [Range(1, int.MaxValue)]
-        public int? ActionCount { get; set; }
-
-        [Display(
-            ResourceType = typeof(OptionsDisplayStrings),
-            Description = nameof(OptionsDisplayStrings.DisplayAttributeDescription_CollectionRuleLimitsOptions_ActionCountSlidingWindowDuration))]
-        [Range(typeof(TimeSpan), CollectionRuleOptionsConstants.ActionCountSlidingWindowDuration_MinValue, CollectionRuleOptionsConstants.ActionCountSlidingWindowDuration_MaxValue)]
-        public TimeSpan? ActionCountSlidingWindowDuration { get; set; }
-
-        [Display(
-            ResourceType = typeof(OptionsDisplayStrings),
-            Description = nameof(OptionsDisplayStrings.DisplayAttributeDescription_CollectionRuleLimitsOptions_RuleDuration))]
-        [Range(typeof(TimeSpan), CollectionRuleOptionsConstants.RuleDuration_MinValue, CollectionRuleOptionsConstants.RuleDuration_MaxValue)]
-        public TimeSpan? RuleDuration { get; set; }
-    }
-}
